requirements.txt
django==3.0.3
djangorestframework==3.11.0
nltk==3.4.5
keras==2.3.1
tensorflow==2.1.0